🎨 Clarified Portal donation naming and Stripe rendering
ref https://github.com/TryGhost/Ghost/pull/26906

The Portal surface is still /support, so the page objects now follow that product naming while the test suite stays donation-focused. The fake checkout HTML also lives in a dedicated renderer module instead of bloating the fake Stripe server.
