+++ e2e/helpers/pages/portal/index.ts
@@ -1,7 +1,7 @@
 export * from './account-page';
-export * from './notification-page';
 export * from './portal-page';
 export * from './sign-in-page';
 export * from './sign-up-page';
 export * from './sign-up-success-page';
+export * from './support-notification-page';
 export * from './support-success-page';
